[phonesim] Add GSMSpecification.xml back. JB#52788
GSMSpecification.xml was left out when build system was converted from
autotools to qmake. Add it back.

Fix packaging mistake with default.xml. It was installed inside extra
directory. To avoid issues with upgrading there is a %pre scriptlet that
removes that directory before installing, otherwise rpm can't replace it
with a file.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>